Moved DimensionsEditor to ScrivitoExtensions
